Version 3.1

+++ src/main/java/co/edu/ims/servicio/res/CategoriaRest.java
@@ -22,7 +22,7 @@ public class CategoriaRest {
     @GET
     @Path("{id}")
     @Produces("application/json")
-    public Categoria buscar(@PathParam("id") Integer pId) {
+    public Categoria buscar(@PathParam("id") Long pId) {
         return em.find(Categoria.class, pId);
     }
